Original change: https://android-review.googlesource.com/c/platform/tools/tradefederation/+/1346643

+++ test_result_interfaces/com/android/tradefed/result/error/ErrorIdentifier.java
@@ -22,5 +22,5 @@ public interface ErrorIdentifier {
     public String name();
 
     /** The unique code identifying the error. */
-    public int code();
+    public long code();
 }
